# docs/swagger.yaml
definitions:
...
  smart-wardrobe-be_internal_modules_subscription_presentation_dto.SetAutoRenewReq:
    properties:
      enabled:
        type: boolean
    required:
    - enabled
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.CategoryRes:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.OutfitItemRes:
    properties:
      id:
        type: string
      layer_order:
        type: integer
      position_x:
        type: number
      position_y:
        type: number
      scale:
        type: number
      wardrobe_item:
        $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes'
    type: object
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.OutfitRes:
    properties:
      cover_image_url:
        type: string
      cover_public_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.OutfitItemRes'
        type: array
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.SaveOutfitItemReq:
    properties:
      layer_order:
        type: integer
      position_x:
        maximum: 1
        minimum: 0
        type: number
      position_y:
        maximum: 1
        minimum: 0
        type: number
      scale:
        maximum: 10
        minimum: 0.1
        type: number
      wardrobe_item_id:
        type: string
    required:
    - layer_order
    - position_x
    - position_y
    - scale
    - wardrobe_item_id
    type: object
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.SaveOutfitReq:
    properties:
      cover_image_url:
        type: string
      cover_public_id:
        maxLength: 255
        type: string
      description:
        type: string
      items:
        items:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.SaveOutfitItemReq'
        type: array
      name:
        maxLength: 255
        type: string
    required:
    - items
    - name
    type: object
...
  smart-wardrobe-be_internal_modules_wardrobe_application_dto.WardrobeItemRes:
    properties:
      category:
        $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.CategoryRes'
      color:
        type: string
      createdAt:
        type: string
      fit:
        type: string
      id:
        type: string
      imagePublicId:
        type: string
      imageUrl:
        type: string
      isLocked:
        type: boolean
      material:
        type: string
      pattern:
        type: string
      seasonality:
        type: string
      status:
        $ref: '#/definitions/smart-wardrobe-be_internal_shared_domain_constants_wardrobestatus.WardrobeItemStatus'
      style:
        type: string
      userId:
        type: string
    type: object
...
  smart-wardrobe-be_internal_shared_domain_constants_wardrobestatus.WardrobeItemStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-comments:
      Failed: AI processing failed
      Processing: AI processing in background
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - AI processing in background
    - AI processing failed
    x-enum-varnames:
    - InWardrobe
    - Selling
    - Sold
    - Processing
    - Failed
  smart-wardrobe-be_internal_shared_presentation.APIResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: API docs for Smart Wardrobe.
  title: Smart Wardrobe API Document
  version: "1.0"
paths:
...
  /api/v1/outfits:
    post:
      consumes:
      - application/json
      description: Lưu bộ phối đồ tự thiết kế cùng danh sách trang phục kèm tọa độ
        kéo thả 2D và layer order.
      parameters:
      - description: Tọa độ 2D và thông tin phối đồ
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.SaveOutfitReq'
      produces:
      - application/json
      responses:
        "201":
          description: Tạo bộ phối đồ thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.OutfitRes'
              type: object
      summary: Tạo bộ phối đồ mới
      tags:
      - Outfits
  /api/v1/outfits/{id}:
    delete:
      description: Xóa bộ phối đồ khỏi bộ sưu tập cá nhân.
      parameters:
      - description: ID bộ phối đồ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa bộ phối đồ thành công
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Xóa bộ phối đồ tự thiết kế
      tags:
      - Outfits
    get:
      description: Trả về chi tiết bộ phối đồ kèm danh sách trang phục đầy đủ và tọa
        độ 2D để render lên canvas.
      parameters:
      - description: ID bộ phối đồ
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chi tiết bộ phối đồ
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.OutfitRes'
              type: object
      summary: Chi tiết bộ phối đồ và tọa độ canvas
      tags:
      - Outfits
    put:
      consumes:
      - application/json
      description: Cập nhật lại thông tin hoặc điều chỉnh tọa độ kéo thả của bộ phối
        đồ.
      parameters:
      - description: ID bộ phối đồ
        in: path
        name: id
        required: true
        type: string
      - description: Thông tin cập nhật
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.SaveOutfitReq'
      produces:
      - application/json
      responses:
        "200":
          description: Cập nhật thành công
          schema:
            allOf:
            - $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/smart-wardrobe-be_internal_modules_wardrobe_application_dto.OutfitRes'
              type: object
      summary: Cập nhật bộ phối đồ
      tags:
      - Outfits
...
  /api/v1/subscriptions/me/auto-renew:
    put:
      consumes:
      - application/json
      description: Thiết lập bật hoặc tắt tính năng tự động gia hạn gói cước qua ví
        nội bộ khi hết hạn
      parameters:
      - description: Trạng thái thiết lập tự động gia hạn
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/smart-wardrobe-be_internal_modules_subscription_presentation_dto.SetAutoRenewReq'
      produces:
      - application/json
      responses:
        "200":
          description: Trạng thái tự động gia hạn mới
          schema:
            $ref: '#/definitions/smart-wardrobe-be_internal_shared_presentation.APIResponse'
      summary: Thiết lập tự động gia hạn gói cước
      tags:
      - Subscription
...
swagger: "2.0"